ROX-29984: rename vm agent to `roxagent`
